% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_RSC.R
\name{cond.RSC}
\alias{cond.RSC}
\title{Condition an operating model for Giant Red Sea Cucumber}
\usage{
cond.RSC(RCMinput, sims = 12, max_F = 0.5, resample = T, silent = T)
}
\arguments{
\item{sims}{Integer or vector of integers - the number of simulations for the operating model (e.g. 96) or the specific vector of simulations for the operating model (e.g. 13:48)}

\item{max_F}{Positive real number - the maximum instantaneous mortality rate in any historical time step}

\item{resample}{Logical - should parameters be drawn from the variance covariance matrix of the MLE fit (T) or just the MLE parameter values (F)}

\item{silent}{Logical - should all RCM messages be repressed (ie not read out onto the console)?}

\item{In}{A list object of class ('In') that includes the operating model parameters (slot 2, class OM) and data (slot 3, class RCMinput) for conditioning}
}
\description{
Uses the RCM model of OpenMSE to fit an operating model to data
}
\examples{
cond.RSC("In.RSC.7F")
}
\seealso{
\link{RCM}
}
\author{
T. Carruthers
}
