% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_RSC.R
\name{MP.RSC}
\alias{MP.RSC}
\title{Management Procedure for Giant Red Sea Cucumber}
\usage{
MP.RSC(
  x,
  Data,
  reps = 1,
  Min.size = NaN,
  Max.size = NaN,
  CEff.Mult = NaN,
  C_I.targ = 1,
  I.targ = 0.5,
  IS.targ = 0,
  IS.yrs = 6,
  IS.fac = 1,
  TAC.calc = "Ratio",
  maxTAC = 5,
  minTAC = 0.1,
  TACdec = 0.2,
  TACinc = 0.2,
  I.enp = 0.25,
  I_freq = c(1, 0, 0, 0, 0),
  rotation = "auto",
  calib_yrs = 20,
  HCR_CP_B = c(0, 0),
  HCR_CP_TAC = c(0, 1),
  curI_2_target = 2,
  DR = 0,
  Fdisc = 0.5
)
}
\arguments{
\item{x}{Positive integer - the simulation number}

\item{Data}{Object of class 'Data'}

\item{reps}{positive integer - not applicable - the number of stochastic draws of advice from which to sample a percentile}

\item{Min.size}{Positive real number - minimum size limit mm}

\item{Max.size}{Positive real number - maximum size limit (NaN is no limit)}

\item{CEff.Mult}{Positive real number (imperfect fraction) the multiplier on current fishing effort (fishing pressure)}

\item{C_I.targ}{Positive real number (imperfect fraction) TAC is calculated TAC(t+1) = I(t) x C(2022)/I(2022) x C_I.targ when TAC.calc = "Ratio"}

\item{I.targ}{Positive real number (imperfect fraction). TAC is reduced / increased to reach I.targ (a fraction of current index) when TAC.calc = "Target"}

\item{IS.targ}{Real number (imperfect fraction) TAC is reduced / increased to reach index target slope (IS.targ) when TAC.calc = "Slope"}

\item{IS.yrs}{Positive integer - the number of years to evalute index slope for the IS rule.}

\item{IS.fac}{Positive real number - sensitivity of the slope - TAC change rule. Lower values are less sensitive. A value of 1 makes changes in proportion to index changes.}

\item{TAC.calc}{Character string. Can be NaN, "Ratio", "Target", "Slope" where TACs are either not constrained by data or set either by index ratio (C_I.targ), index target (I.targ) or index slope target (IS.targ)}

\item{maxTAC}{Positive real number (imperfect fraction) - annual catches cannot exceed current catches muliplied by this factor}

\item{minTAC}{Positive real number (imperfect fraction) - annual catches cannot be lower than current catches muliplied by this factor}

\item{TACdec}{Positive real number (fraction) - downward TAC changes cannot exceed this fraction (e.g. 0.2 is maximum decline of 20 per cent among management cycles)}

\item{TACinc}{Positive real number (fraction) - upward TAC changes cannot exceed this fraction (e.g. 0.1 is maximum increase of 10 per cent among management cycles)}

\item{I.enp}{Positive real number (fraction) the parameter controlling effective number of parameters for the polynomial smoother on the indices. npar = ny * I.enp so higher values mean more smoother parameters and less smoothing}

\item{I_freq}{Vector of positive integers - how frequently do you collect the indices (0 = do not collect, 1 = every year, 2 = every other year, 3 = once every three years, ...).}

\item{rotation}{Vector proyears long of catch on/off (1 and 0, respectively). If a single value is specified (1, 2,..) the pattern alternates on once every year (1), every other year (2) etc. The default is "auto" where the MP tried to detect the most recent pattern}

\item{calib_yrs}{Positive integer - how many of the recent years are used to calibrate Index to catch ratio for TAC based MPs}

\item{HCR_CP_B}{A positive numeric vector two positions long of biomass control points (x axis) relative to recent index / curI_2_target for a hockey stick harvest control rule c(0,0) essentially has no control rule if HCR_CP_TAC = 1}

\item{HCR_CP_TAC}{A positive numeric vector two positions long that is the fraction of the recommended TAC taken below control point 1 and above control point 2 (y axis adjustment of the harvest control rule)}

\item{curI_2_target}{A positive real number that is the level of the current index (recent historical year) relative to the target biomass level (e.g. BMSY) 2 implies recent indices are at twice target levels (underexploited)}

\item{DR}{Fraction - the discard rate}

\item{Fdisc}{The fraction of discarded individuals that die}
}
\description{
A modifiable management procedure for Giant Red Sea Cucumber that allows for adjustments to size limits, current effort and TAC control via indices
}
\examples{
MP.RSC(Simulation_1) # apply to a generic simulated dataset from openMSE
}
\author{
T. Carruthers
}
